requirements.txt
python-telegram-bot
schedule
mnk-persian-words